requirements.txt
streamlit
PyMuPDF
sentence-transformers
faiss-cpu
transformers
torch
python-dotenv